move methods to correct interface, deprecate private interface
