+++ lib/private/user/interface.php
@@ -25,62 +25,8 @@
  *
  */
 
-interface OC_User_Interface {
-
-	/**
-	 * Check if backend implements actions
-	 * @param int $actions bitwise-or'ed actions
-	 * @return boolean
-	 *
-	 * Returns the supported actions as int to be
-	 * compared with \OC_User_Backend::CREATE_USER etc.
-	 */
-	public function implementsActions($actions);
-
-	/**
-	 * delete a user
-	 * @param string $uid The username of the user to delete
-	 * @return bool
-	 */
-	public function deleteUser($uid);
-
-	/**
-	 * Get a list of all users
-	 *
-	 * @param string $search
-	 * @param null|int $limit
-	 * @param null|int $offset
-	 * @return string[] an array of all uids
-	 */
-	public function getUsers($search = '', $limit = null, $offset = null);
-
-	/**
-	 * check if a user exists
-	 * @param string $uid the username
-	 * @return boolean
-	 */
-	public function userExists($uid);
-
-	/**
-	 * get display name of the user
-	 * @param string $uid user ID of the user
-	 * @return string display name
-	 */
-	public function getDisplayName($uid);
-
-	/**
-	 * Get a list of all display names and user ids.
-	 *
-	 * @param string $search
-	 * @param string|null $limit
-	 * @param string|null $offset
-	 * @return array an array of all displayNames (value) and the corresponding uids (key)
-	 */
-	public function getDisplayNames($search = '', $limit = null, $offset = null);
-
-	/**
-	 * Check if a user list is available or not
-	 * @return boolean if users can be listed or not
-	 */
-	public function hasUserListings();
-}
+/**
+ * Interface OC_User_Interface
+ * @deprecated use the public \OCP\UserInterface instead
+ */
+interface OC_User_Interface extends \OCP\UserInterface {}
