+++ lib/public/groupinterface.php
@@ -37,4 +37,70 @@ namespace OCP;
  * @package OCP
  * @since 4.5.0
  */
-interface GroupInterface extends \OC_Group_Interface {}
+interface GroupInterface {
+
+	/**
+	 * Check if backend implements actions
+	 * @param int $actions bitwise-or'ed actions
+	 * @return boolean
+	 * @since 4.5.0
+	 *
+	 * Returns the supported actions as int to be
+	 * compared with \OC_Group_Backend::CREATE_GROUP etc.
+	 */
+	public function implementsActions($actions);
+
+	/**
+	 * is user in group?
+	 * @param string $uid uid of the user
+	 * @param string $gid gid of the group
+	 * @return bool
+	 * @since 4.5.0
+	 *
+	 * Checks whether the user is member of a group or not.
+	 */
+	public function inGroup($uid, $gid);
+
+	/**
+	 * Get all groups a user belongs to
+	 * @param string $uid Name of the user
+	 * @return array an array of group names
+	 * @since 4.5.0
+	 *
+	 * This function fetches all groups a user belongs to. It does not check
+	 * if the user exists at all.
+	 */
+	public function getUserGroups($uid);
+
+	/**
+	 * get a list of all groups
+	 * @param string $search
+	 * @param int $limit
+	 * @param int $offset
+	 * @return array an array of group names
+	 * @since 4.5.0
+	 *
+	 * Returns a list with all groups
+	 */
+	public function getGroups($search = '', $limit = -1, $offset = 0);
+
+	/**
+	 * check if a group exists
+	 * @param string $gid
+	 * @return bool
+	 * @since 4.5.0
+	 */
+	public function groupExists($gid);
+
+	/**
+	 * get a list of all users in a group
+	 * @param string $gid
+	 * @param string $search
+	 * @param int $limit
+	 * @param int $offset
+	 * @return array an array of user ids
+	 * @since 4.5.0
+	 */
+	public function usersInGroup($gid, $search = '', $limit = -1, $offset = 0);
+
+}
